feat(#17): validate column configuration
