+++ internal/csv2json/helper_test.go
@@ -62,3 +62,61 @@ func TestSetValue(t *testing.T) {
 		})
 	}
 }
+
+// TestSetValueInternal tests the setValueInternal function which recursively creates and maps nested dictionaries.
+func TestSetValueInternal(t *testing.T) {
+	tests := []struct {
+		name      string
+		hierarchy []string
+		value     any
+		inside    map[string]any
+		want      any
+	}{
+		{
+			name:      "single level",
+			hierarchy: []string{"key"},
+			value:     "value",
+			inside:    map[string]any{},
+			want:      "value",
+		},
+		{
+			name:      "two levels",
+			hierarchy: []string{"parent", "child"},
+			value:     "value",
+			inside:    map[string]any{},
+			want:      map[string]any{"child": "value"},
+		},
+		{
+			name:      "three levels",
+			hierarchy: []string{"level1", "level2", "level3"},
+			value:     "value",
+			inside:    map[string]any{},
+			want:      map[string]any{"level2": map[string]any{"level3": "value"}},
+		},
+		{
+			name:      "existing nested data",
+			hierarchy: []string{"parent", "child1", "grandchild"},
+			value:     "value",
+			inside: map[string]any{
+				"parent": map[string]any{
+					"child2": "existing",
+				},
+			},
+			want: map[string]any{
+				"child1": map[string]any{
+					"grandchild": "value",
+				},
+				"child2": "existing",
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := setValueInternal(tt.hierarchy, tt.value, tt.inside)
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("setValueInternal() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
